Create 1.PGMovies.js
A .js file containing a class Movie. Each instance of the Movie class has the properties title, studio and rating. The default rating is set as "PG". A method getPG takes an array of base type Movie and returns a new array of the movies which have a "PG" rating.