Create 2.CircleClass.js
A .js file conatining a class Circle. The constructor has the properties radius and color, where radius is set to 1 and the color is set to "red" by default. It contains various methods that help to get the radius, color, set the radius, set the color, get the circumference and get the area of the circle.